backend/requirements.txt
fastapi
uvicorn
motor
pydantic
bcrypt
python-jose
httpx
python-multipart